+++ src/pages/Howto/Content/Howto/Step/Step.tsx
@@ -4,7 +4,7 @@ import ReactPlayer from 'react-player'
 import { Box, Card, Text } from 'theme-ui'
 import { Flex } from 'theme-ui'
 import Heading from 'src/components/Heading'
-import ImageGallery from 'src/components/ImageGallery'
+import { ImageGallery } from 'src/components/ImageGallery/ImageGallery'
 import type { IHowtoStep } from 'src/models/howto.models'
 import type { IUploadedFileMeta } from 'src/stores/storage'
 import { capitalizeFirstLetter } from 'src/utils/helpers'
