+++ src/pages/Research/Content/ResearchUpdate.tsx
@@ -6,7 +6,7 @@ import { Box, Card, Text } from 'theme-ui'
 import { Button } from 'oa-components'
 import { Flex } from 'theme-ui'
 import Heading from 'src/components/Heading'
-import ImageGallery from 'src/components/ImageGallery'
+import { ImageGallery } from 'src/components/ImageGallery/ImageGallery'
 import { Link } from 'theme-ui'
 import type { IResearch } from 'src/models/research.models'
 import type { IUploadedFileMeta } from 'src/stores/storage'
